+++ flang-rt/lib/runtime/findloc.h
@@ -1,4 +1,4 @@
-//===-- lib/runtime/findloc.cpp ---------------------------------*- C++ -*-===//
+//===-- lib/runtime/findloc.h -----------------------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -8,6 +8,9 @@
 
 // Implements FINDLOC for all required operand types and shapes and result
 // integer kinds.
+
+#ifndef FLANG_RT_RUNTIME_FINDLOC_H_
+#define FLANG_RT_RUNTIME_FINDLOC_H_
 
 #include "flang-rt/runtime/reduction-templates.h"
 #include "flang/Runtime/character.h"
@@ -243,7 +246,7 @@ template <int KIND> struct CharacterFindlocHelper {
   }
 };
 
-static RT_API_ATTRS void LogicalFindlocHelper(Descriptor &result,
+inline RT_API_ATTRS void LogicalFindlocHelper(Descriptor &result,
     const Descriptor &x, const Descriptor &target, int kind,
     const Descriptor *mask, bool back, Terminator &terminator) {
   using Accumulator = LocationAccumulator<LogicalEquivalence>;
@@ -252,68 +255,6 @@ static RT_API_ATTRS void LogicalFindlocHelper(Descriptor &result,
   ApplyIntegerKind<LocationResultHelper<Accumulator>::template Functor, void>(
       kind, terminator, accumulator, result);
 }
-
-extern "C" {
-RT_EXT_API_GROUP_BEGIN
-
-void RTDEF(Findloc)(Descriptor &result, const Descriptor &x,
-    const Descriptor &target, int kind, const char *source, int line,
-    const Descriptor *mask, bool back) {
-  int rank{x.rank()};
-  SubscriptValue extent[1]{rank};
-  result.Establish(TypeCategory::Integer, kind, nullptr, 1, extent,
-      CFI_attribute_allocatable);
-  result.GetDimension(0).SetBounds(1, extent[0]);
-  Terminator terminator{source, line};
-  if (int stat{result.Allocate(kNoAsyncObject)}) {
-    terminator.Crash(
-        "FINDLOC: could not allocate memory for result; STAT=%d", stat);
-  }
-  CheckIntegerKind(terminator, kind, "FINDLOC");
-  auto xType{x.type().GetCategoryAndKind()};
-  auto targetType{target.type().GetCategoryAndKind()};
-  RUNTIME_CHECK(terminator, xType.has_value() && targetType.has_value());
-  switch (xType->first) {
-  case TypeCategory::Integer:
-  case TypeCategory::Unsigned:
-    ApplyIntegerKind<
-        TotalNumericFindlocSource<TypeCategory::Integer>::template Functor,
-        void>(xType->second, terminator, targetType->first, targetType->second,
-        result, x, target, kind, 0, mask, back, terminator);
-    break;
-  case TypeCategory::Real:
-    ApplyFloatingPointKind<
-        TotalNumericFindlocSource<TypeCategory::Real>::template Functor, void>(
-        xType->second, terminator, targetType->first, targetType->second,
-        result, x, target, kind, 0, mask, back, terminator);
-    break;
-  case TypeCategory::Complex:
-    ApplyFloatingPointKind<
-        TotalNumericFindlocSource<TypeCategory::Complex>::template Functor,
-        void>(xType->second, terminator, targetType->first, targetType->second,
-        result, x, target, kind, 0, mask, back, terminator);
-    break;
-  case TypeCategory::Character:
-    RUNTIME_CHECK(terminator,
-        targetType->first == TypeCategory::Character &&
-            targetType->second == xType->second);
-    ApplyCharacterKind<CharacterFindlocHelper, void>(xType->second, terminator,
-        result, x, target, kind, mask, back, terminator);
-    break;
-  case TypeCategory::Logical:
-    RUNTIME_CHECK(terminator, targetType->first == TypeCategory::Logical);
-    LogicalFindlocHelper(result, x, target, kind, mask, back, terminator);
-    break;
-  default:
-    terminator.Crash(
-        "FINDLOC: bad data type code (%d) for array", x.type().raw());
-  }
-}
-
-RT_EXT_API_GROUP_END
-} // extern "C"
-
-// FINDLOC with DIM=
 
 template <TypeCategory CAT> struct PartialNumericFindlocSource {
   template <int KIND> struct Functor {
@@ -343,7 +284,7 @@ template <int KIND> struct PartialCharacterFindlocHelper {
   }
 };
 
-static RT_API_ATTRS void PartialLogicalFindlocHelper(Descriptor &result,
+inline RT_API_ATTRS void PartialLogicalFindlocHelper(Descriptor &result,
     const Descriptor &x, const Descriptor &target, int kind, int dim,
     const Descriptor *mask, bool back, Terminator &terminator) {
   using Accumulator = LocationAccumulator<LogicalEquivalence>;
@@ -353,55 +294,6 @@ static RT_API_ATTRS void PartialLogicalFindlocHelper(Descriptor &result,
       accumulator);
 }
 
-extern "C" {
-RT_EXT_API_GROUP_BEGIN
-
-void RTDEF(FindlocDim)(Descriptor &result, const Descriptor &x,
-    const Descriptor &target, int kind, int dim, const char *source, int line,
-    const Descriptor *mask, bool back) {
-  Terminator terminator{source, line};
-  CheckIntegerKind(terminator, kind, "FINDLOC");
-  auto xType{x.type().GetCategoryAndKind()};
-  auto targetType{target.type().GetCategoryAndKind()};
-  RUNTIME_CHECK(terminator, xType.has_value() && targetType.has_value());
-  switch (xType->first) {
-  case TypeCategory::Integer:
-  case TypeCategory::Unsigned:
-    ApplyIntegerKind<
-        PartialNumericFindlocSource<TypeCategory::Integer>::template Functor,
-        void>(xType->second, terminator, targetType->first, targetType->second,
-        result, x, target, kind, dim, mask, back, terminator);
-    break;
-  case TypeCategory::Real:
-    ApplyFloatingPointKind<
-        PartialNumericFindlocSource<TypeCategory::Real>::template Functor,
-        void>(xType->second, terminator, targetType->first, targetType->second,
-        result, x, target, kind, dim, mask, back, terminator);
-    break;
-  case TypeCategory::Complex:
-    ApplyFloatingPointKind<
-        PartialNumericFindlocSource<TypeCategory::Complex>::template Functor,
-        void>(xType->second, terminator, targetType->first, targetType->second,
-        result, x, target, kind, dim, mask, back, terminator);
-    break;
-  case TypeCategory::Character:
-    RUNTIME_CHECK(terminator,
-        targetType->first == TypeCategory::Character &&
-            targetType->second == xType->second);
-    ApplyCharacterKind<PartialCharacterFindlocHelper, void>(xType->second,
-        terminator, result, x, target, kind, dim, mask, back, terminator);
-    break;
-  case TypeCategory::Logical:
-    RUNTIME_CHECK(terminator, targetType->first == TypeCategory::Logical);
-    PartialLogicalFindlocHelper(
-        result, x, target, kind, dim, mask, back, terminator);
-    break;
-  default:
-    terminator.Crash(
-        "FINDLOC: bad data type code (%d) for array", x.type().raw());
-  }
-}
-
-RT_EXT_API_GROUP_END
-} // extern "C"
 } // namespace Fortran::runtime
+
+#endif // FLANG_RT_RUNTIME_FINDLOC_H_
