+++ flang-rt/lib/runtime/matmul.h
@@ -1,4 +1,4 @@
-//===-- lib/runtime/matmul.cpp ----------------------------------*- C++ -*-===//
+//===-- lib/runtime/matmul.h ------------------------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -19,13 +19,16 @@
 //
 // Places where BLAS routines could be called are marked as TODO items.
 
-#include "flang/Runtime/matmul.h"
+#ifndef FLANG_RT_RUNTIME_MATMUL_H_
+#define FLANG_RT_RUNTIME_MATMUL_H_
+
 #include "flang-rt/runtime/descriptor.h"
 #include "flang-rt/runtime/terminator.h"
 #include "flang-rt/runtime/tools.h"
 #include "flang/Common/optional.h"
 #include "flang/Runtime/c-or-cpp.h"
 #include "flang/Runtime/cpp-type.h"
+#include "flang/Runtime/matmul.h"
 #include <cstring>
 
 namespace {
@@ -56,23 +59,6 @@ private:
   Result sum_{};
 };
 
-// Contiguous numeric matrix*matrix multiplication
-//   matrix(rows,n) * matrix(n,cols) -> matrix(rows,cols)
-// Straightforward algorithm:
-//   DO 1 I = 1, NROWS
-//    DO 1 J = 1, NCOLS
-//     RES(I,J) = 0
-//     DO 1 K = 1, N
-//   1  RES(I,J) = RES(I,J) + X(I,K)*Y(K,J)
-// With loop distribution and transposition to avoid the inner sum
-// reduction and to avoid non-unit strides:
-//   DO 1 I = 1, NROWS
-//    DO 1 J = 1, NCOLS
-//   1 RES(I,J) = 0
-//   DO 2 K = 1, N
-//    DO 2 J = 1, NCOLS
-//     DO 2 I = 1, NROWS
-//   2  RES(I,J) = RES(I,J) + X(I,K)*Y(K,J) ! loop-invariant last term
 template <TypeCategory RCAT, int RKIND, typename XT, typename YT,
     bool X_HAS_STRIDED_COLUMNS, bool Y_HAS_STRIDED_COLUMNS>
 inline RT_API_ATTRS void MatrixTimesMatrix(
@@ -132,20 +118,6 @@ inline RT_API_ATTRS void MatrixTimesMatrixHelper(
   }
 }
 
-// Contiguous numeric matrix*vector multiplication
-//   matrix(rows,n) * column vector(n) -> column vector(rows)
-// Straightforward algorithm:
-//   DO 1 J = 1, NROWS
-//    RES(J) = 0
-//    DO 1 K = 1, N
-//   1 RES(J) = RES(J) + X(J,K)*Y(K)
-// With loop distribution and transposition to avoid the inner
-// sum reduction and to avoid non-unit strides:
-//   DO 1 J = 1, NROWS
-//   1 RES(J) = 0
-//   DO 2 K = 1, N
-//    DO 2 J = 1, NROWS
-//   2 RES(J) = RES(J) + X(J,K)*Y(K)
 template <TypeCategory RCAT, int RKIND, typename XT, typename YT,
     bool X_HAS_STRIDED_COLUMNS>
 inline RT_API_ATTRS void MatrixTimesVector(
@@ -182,20 +154,6 @@ inline RT_API_ATTRS void MatrixTimesVectorHelper(
   }
 }
 
-// Contiguous numeric vector*matrix multiplication
-//   row vector(n) * matrix(n,cols) -> row vector(cols)
-// Straightforward algorithm:
-//   DO 1 J = 1, NCOLS
-//    RES(J) = 0
-//    DO 1 K = 1, N
-//   1 RES(J) = RES(J) + X(K)*Y(K,J)
-// With loop distribution and transposition to avoid the inner
-// sum reduction and one non-unit stride (the other remains):
-//   DO 1 J = 1, NCOLS
-//   1 RES(J) = 0
-//   DO 2 K = 1, N
-//    DO 2 J = 1, NCOLS
-//   2 RES(J) = RES(J) + X(K)*Y(K,J)
 template <TypeCategory RCAT, int RKIND, typename XT, typename YT,
     bool Y_HAS_STRIDED_COLUMNS>
 inline RT_API_ATTRS void VectorTimesMatrix(
@@ -234,7 +192,6 @@ inline RT_API_ATTRS void VectorTimesMatrixHelper(
   }
 }
 
-// Implements an instance of MATMUL for given argument types.
 template <TypeCategory RCAT, int RKIND, typename XT, typename YT>
 static inline RT_API_ATTRS void DoMatmul(Descriptor &result,
     const Descriptor &x, const Descriptor &y, Terminator &terminator,
@@ -268,8 +225,6 @@ static inline RT_API_ATTRS void DoMatmul(Descriptor &result,
   }
   SubscriptValue n{x.GetDimension(xRank - 1).Extent()};
   if (n != y.GetDimension(0).Extent()) {
-    // At this point, we know that there's a shape error.  There are three
-    // possibilities, x is rank 1, y is rank 1, or both are rank 2.
     if (xRank == 1) {
       terminator.Crash("MATMUL: unacceptable operand shapes (%jd, %jdx%jd)",
           static_cast<std::intmax_t>(n),
@@ -294,11 +249,8 @@ static inline RT_API_ATTRS void DoMatmul(Descriptor &result,
   if constexpr (RCAT != TypeCategory::Logical) {
     if (x.IsContiguous(1) && y.IsContiguous(1) &&
         (isAllocating || result.IsContiguous())) {
-      // Contiguous numeric matrices (maybe with columns
-      // separated by a stride).
       Fortran::common::optional<std::size_t> xColumnByteStride;
       if (!x.IsContiguous()) {
-        // X's columns are strided.
         SubscriptValue xAt[2]{};
         x.GetLowerBounds(xAt);
         xAt[1]++;
@@ -306,16 +258,11 @@ static inline RT_API_ATTRS void DoMatmul(Descriptor &result,
       }
       Fortran::common::optional<std::size_t> yColumnByteStride;
       if (!y.IsContiguous()) {
-        // Y's columns are strided.
         SubscriptValue yAt[2]{};
         y.GetLowerBounds(yAt);
         yAt[1]++;
         yColumnByteStride = y.SubscriptsToByteOffset(yAt);
       }
-      // Note that BLAS GEMM can be used for the strided
-      // columns by setting proper leading dimension size.
-      // This implies that the column stride is divisible
-      // by the element size, which is usually true.
       if (resRank == 2) { // M*M -> M
         if (std::is_same_v<XT, YT>) {
           if constexpr (std::is_same_v<XT, float>) {
@@ -369,7 +316,6 @@ static inline RT_API_ATTRS void DoMatmul(Descriptor &result,
       }
     }
   }
-  // General algorithms for LOGICAL and noncontiguity
   SubscriptValue xAt[2], yAt[2], resAt[2];
   x.GetLowerBounds(xAt);
   y.GetLowerBounds(yAt);
@@ -449,10 +395,6 @@ struct MatmulHelper {
 };
 } // namespace
 
-namespace Fortran::runtime {
-extern "C" {
-RT_EXT_API_GROUP_BEGIN
-
 #define MATMUL_INSTANCE(XCAT, XKIND, YCAT, YKIND) \
   void RTDEF(Matmul##XCAT##XKIND##YCAT##YKIND)(Descriptor & result, \
       const Descriptor &x, const Descriptor &y, const char *sourceFile, \
@@ -469,10 +411,4 @@ RT_EXT_API_GROUP_BEGIN
         result, x, y, sourceFile, line, false); \
   }
 
-#define MATMUL_FORCE_ALL_TYPES 0
-
-#include "flang/Runtime/matmul-instances.inc"
-
-RT_EXT_API_GROUP_END
-} // extern "C"
-} // namespace Fortran::runtime
+#endif // FLANG_RT_RUNTIME_MATMUL_H_
